style: simplify file path
