+++ api-server/src/main/java/nexters/dividend/apiserver/DividendApiServerApplication.java
@@ -1,0 +1,13 @@
+package nexters.dividend.apiserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DividendApiServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DividendApiServerApplication.class, args);
+	}
+
+}
