+++ api-server/src/test/java/nexters/dividend/apiserver/DividendApiServerApplicationTests.java
@@ -1,0 +1,13 @@
+package nexters.dividend.apiserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DividendApiServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
