+++ batch/src/main/java/nexters/dividend/batch/DividendBatchApplication.java
@@ -1,0 +1,13 @@
+package nexters.dividend.batch;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DividendBatchApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DividendBatchApplication.class, args);
+    }
+
+}
