+++ batch/src/test/java/nexters/dividend/batch/DividendBatchApplicationTests.java
@@ -1,0 +1,13 @@
+package nexters.dividend.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DividendBatchApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
