+++ core/src/main/java/nexters/dividend/core/DividendCoreApplication.java
@@ -1,0 +1,13 @@
+package nexters.dividend.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DividendCoreApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DividendCoreApplication.class, args);
+    }
+
+}
