+++ core/src/test/java/nexters/dividend/core/DividendCoreApplicationTests.java
@@ -1,0 +1,13 @@
+package nexters.dividend.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DividendCoreApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
