--- dividend-api-server/src/main/java/nexters/dividendapiserver/DividendApiServerApplication.java
@@ -1,13 +1,0 @@
-package nexters.dividendapiserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DividendApiServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DividendApiServerApplication.class, args);
-	}
-
-}
