--- dividend-api-server/src/test/java/nexters/dividendapiserver/DividendApiServerApplicationTests.java
@@ -1,13 +1,0 @@
-package nexters.dividendapiserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DividendApiServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
