--- dividend-batch/src/main/java/nexters/dividendbatch/DividendBatchApplication.java
@@ -1,13 +1,0 @@
-package nexters.dividendbatch;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DividendBatchApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DividendBatchApplication.class, args);
-    }
-
-}
