--- dividend-batch/src/test/java/nexters/dividendbatch/DividendBatchApplicationTests.java
@@ -1,13 +1,0 @@
-package nexters.dividendbatch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DividendBatchApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
