--- dividend-core/src/main/java/nexters/dividendcore/DividendCoreApplication.java
@@ -1,13 +1,0 @@
-package nexters.dividendcore;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DividendCoreApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DividendCoreApplication.class, args);
-    }
-
-}
