--- dividend-core/src/test/java/nexters/dividendcore/DividendCoreApplicationTests.java
@@ -1,13 +1,0 @@
-package nexters.dividendcore;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DividendCoreApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
