+++ domain/src/main/java/nexters/dividend/domain/DomainApplication.java
@@ -1,0 +1,13 @@
+package nexters.dividend.domain;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DomainApplication.class, args);
+    }
+
+}
