+++ domain/src/main/java/nexters/dividend/domain/stock/StockRepository.java
@@ -1,0 +1,9 @@
+package nexters.dividend.domain.stock;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface StockRepository extends JpaRepository<Stock, UUID> {
+
+}
