--- domain/src/main/java/nexters/domain/DomainApplication.java
@@ -1,13 +1,0 @@
-package nexters.domain;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DomainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DomainApplication.class, args);
-    }
-
-}
