--- domain/src/main/java/nexters/domain/stock/StockRepository.java
@@ -1,9 +1,0 @@
-package nexters.domain.stock;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface StockRepository extends JpaRepository<Stock, UUID> {
-
-}
