+++ domain/src/test/java/nexters/dividend/domain/DomainApplicationTests.java
@@ -1,0 +1,13 @@
+package nexters.dividend.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
