--- domain/src/test/java/nexters/domain/DomainApplicationTests.java
@@ -1,13 +1,0 @@
-package nexters.domain;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DomainApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
